+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/default_dl_file_type.ftl
@@ -1,9 +1,9 @@
-${dataFactory.toInsertSQL(dataFactory.newDLFileEntryTypeModel())}
+${dataFactory.toInsertSQL(dLDataFactory.newDLFileEntryTypeModel())}
 
-<#assign defaultDLDDMStructureModel = dataFactory.newDefaultDLDDMStructureModel() />
+<#assign defaultDLDDMStructureModel = dLDataFactory.newDefaultDLDDMStructureModel() />
 
 <@insertDDMStructure
-	_ddmStructureLayoutModel=dataFactory.newDefaultDLDDMStructureLayoutModel()
+	_ddmStructureLayoutModel=dLDataFactory.newDefaultDLDDMStructureLayoutModel()
 	_ddmStructureModel=defaultDLDDMStructureModel
-	_ddmStructureVersionModel=dataFactory.newDefaultDLDDMStructureVersionModel(defaultDLDDMStructureModel)
+	_ddmStructureVersionModel=dLDataFactory.newDefaultDLDDMStructureVersionModel(defaultDLDDMStructureModel)
 />
